requirements.txt
requests~=2.27.1
python-dotenv~=0.20.0
wonderwords~=2.2.0
essential-generators~=1.0
Pillow~=9.1.0
tweepy~=4.8.0